refactor(boundary): Move SSL and writable output contract
Move SSL into network.crypta.node.runtime so future crypto-keys extraction work no longer carries config glue inside network.crypta.crypt.

Move WritableToDataOutputStream into foundation-support and register its owned output pattern so stale aggregated root outputs are pruned correctly.

Update affected production and test imports to use the relocated types without changing behavior.
