+++ foundation-support/src/main/java/network/crypta/io/WritableToDataOutputStream.java
