+++ src/main/java/network/crypta/clients/http/SimpleToadletServer.java
@@ -25,7 +25,6 @@ import network.crypta.config.NodeNeedRestartException;
 import network.crypta.config.Option;
 import network.crypta.config.StringCallback;
 import network.crypta.config.SubConfig;
-import network.crypta.crypt.SSL;
 import network.crypta.io.AllowedHosts;
 import network.crypta.io.NetworkInterface;
 import network.crypta.io.SSLNetworkInterface;
@@ -33,6 +32,7 @@ import network.crypta.keys.FreenetURI;
 import network.crypta.l10n.NodeL10n;
 import network.crypta.node.NodeClientCore;
 import network.crypta.node.PrioRunnable;
+import network.crypta.node.runtime.SSL;
 import network.crypta.node.useralerts.UserAlertManager;
 import network.crypta.runtime.spi.RandomnessPort;
 import network.crypta.runtime.spi.RuntimePorts;
