+++ src/main/java/network/crypta/clients/http/ToadletContextImpl.java
@@ -27,8 +27,8 @@ import java.util.TimeZone;
 import java.util.concurrent.atomic.AtomicReference;
 import network.crypta.clients.http.FProxyFetchInProgress.REFILTER_POLICY;
 import network.crypta.clients.http.bookmark.BookmarkManager;
-import network.crypta.crypt.SSL;
 import network.crypta.l10n.NodeL10n;
+import network.crypta.node.runtime.SSL;
 import network.crypta.node.useralerts.UserAlertManager;
 import network.crypta.support.HTMLEncoder;
 import network.crypta.support.HTMLNode;
