+++ src/main/java/network/crypta/node/NodeStarter.java
@@ -23,12 +23,12 @@ import network.crypta.config.SubConfig;
 import network.crypta.crypt.CryptoRandoms;
 import network.crypta.crypt.JceLoader;
 import network.crypta.crypt.RandomSource;
-import network.crypta.crypt.SSL;
 import network.crypta.crypt.Yarrow;
 import network.crypta.fs.AppDirs;
 import network.crypta.fs.AppEnv;
 import network.crypta.fs.Resolved;
 import network.crypta.fs.ServiceDirs;
+import network.crypta.node.runtime.SSL;
 import network.crypta.support.JVMVersion;
 import network.crypta.support.Logging;
 import network.crypta.support.PooledExecutor;
