+++ src/test/java/network/crypta/node/TextModeClientInterfaceServerTest.java
@@ -10,8 +10,8 @@ import network.crypta.config.Config;
 import network.crypta.config.InvalidConfigValueException;
 import network.crypta.config.SubConfig;
 import network.crypta.crypt.RandomSource;
-import network.crypta.crypt.SSL;
 import network.crypta.io.NetworkInterface;
+import network.crypta.node.runtime.SSL;
 import network.crypta.node.subsystem.NodeBootstrap;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
 import network.crypta.support.PriorityAwareExecutor;
